Tweaks to girepository vapi to get InvokeError and Arguments working.
In the end it required a #define in a special header, because I was not
able to persuade valac to generate correct domain name for
Introspection.InvokeError (it still generated G_IINVOKE_ERROR). Abused the
header to silence the warnings about g_base_info_ref/g_base_info_unref
being called with incompatible pointer type (that's probably vala bug).

Signed-off-by: Jan Hudec <bulb@ucw.cz>