Actually start the test.
The option parsing is now two stage, because some options are passed to
g_test_init. Because it makes the critical warning in
gobject-introspection on unresolved symbol (it's a bug that it's
a critical warning) and also because it would handle --help, we have to
parse our options first and it's options later. That makes it hard to
handle positional options. Unfortunately it seems g_test_run does not know
to skip test cases by pattern, so we'll have to do it ourselves and we'll
have to do it by option, because of the argument difficulty. Anyway.

Signed-off-by: Jan Hudec <bulb@ucw.cz>