Add support for reading vapi files.
A .vapi file can now be specified instead of a .gir file and it will be
parsed using libvala and all the tests will be executed. Only synchronous
tests are supported so far.

Tests work except for a problem with the async operation in
test_wait_async_fail not being properly cancelled and crashing the next
test.

Signed-off-by: Jan Hudec <bulb@ucw.cz>