Enable testing wait_for_cancellable_async.
Lambdas in async method were finally fixed upstream, so I could write
a cancellable async method to test waiting for it. There is a bug related
to connecting lambdas to signals (no way to disconnect them), but it does
not cause problem for the test case as the emitter is than destroyed
anyway.

Signed-off-by: Jan Hudec <bulb@ucw.cz>