Workaround overzealous check in GLib 2.22
Glib 2.22 checks in g_simple_async_result_finish that it's being called in
correct context. Unfortunately that check complains with critical (which
is fatal under test suite) that the source is destroyed if we remove it
before that. Since g_simple_async_result_finish is called in the return
statement of the async method, we have to make sure the source is removed
after the function finishes. I always thought it was legal to remove
a source in it's dispatch. Anyway.

Signed-off-by: Jan Hudec <bulb@ucw.cz>