Bump version number to 0.1 and API version to 1.0
API version will likely stay that forever, it makes only sense to change
that if the api change is completely incompatible.

The SO version currently remains at 0.0 and will from now on be
incremented whenever something changes.

The build files were changed so that the version numbers don't appear
anywhere except at the begining of main wscript to ease maintenance.

Signed-off-by: Jan Hudec <bulb@ucw.cz>